Refactor scroll into view logic to component manager
